wip add sort order

+++ server/queries/story-content-queries.js
@@ -166,15 +166,100 @@ async function insertNewChoice(
     };
   }
 
+  const sortOrderArr = (
+    await pbPool.query(
+      "SELECT sort_order FROM choices WHERE from_passage_id = $1 ORDER BY sort_order ASC",
+      [from_passage_id],
+    )
+  ).rows.map((row) => row.sort_order);
+
+  let newSortOrder = 0;
+  if (sortOrderArr.length > 0 && sortOrderArr.includes(0)) {
+    for (let i = 0; i < sortOrderArr.length; i++) {
+      if (sortOrderArr[i + 1] - sortOrderArr[i] > 1) {
+        newSortOrder =
+          sortOrderArr[i] === 0 ? sortOrderArr[i] : sortOrderArr[i] + 1;
+        break;
+      } else if (i === sortOrderArr.length - 1) {
+        newSortOrder = sortOrderArr[i] + 1;
+      }
+    }
+  }
+
   const SQL = `
-    INSERT INTO choices (label, from_passage_id, to_passage_id) VALUES ($1, $2, $3) RETURNING *
+    INSERT INTO choices (label, from_passage_id, to_passage_id, sort_order) VALUES ($1, $2, $3, $4) RETURNING *
   `;
 
   const { rows } = await pbPool.query(SQL, [
     label,
     from_passage_id,
     to_passage_id,
+    newSortOrder,
   ]);
+  return rows;
+}
+
+async function updateChoiceSortOrderById(
+  user_id,
+  story_id,
+  choice_id,
+  sort_order,
+) {
+  const isOwner = await storyOwnerCheck(user_id, story_id);
+  if (!isOwner) return null;
+
+  const fromPassageIdResults = await pbPool.query(
+    "SELECT from_passage_id FROM choices WHERE id = $1",
+    [choice_id],
+  );
+
+  if (!fromPassageIdResults.rowCount > 0) {
+    return {
+      queryError: "choice not found",
+    };
+  }
+
+  const fromPassageId = fromPassageIdResults.rows[0].from_passage_id;
+
+  const currentSort = (
+    await pbPool.query("SELECT sort_order FROM choices WHERE id = $1", [
+      choice_id,
+    ])
+  ).rows[0].sort_order;
+
+  const sortOrderArr = (
+    await pbPool.query(
+      "SELECT sort_order FROM choices WHERE from_passage_id = $1",
+      [fromPassageId],
+    )
+  ).rows.map((row) => row.sort_order);
+
+  const existingChoice = await pbPool.query(
+    "SELECT id FROM choices WHERE from_passage_id = $1 AND sort_order = $2",
+    [fromPassageId, sort_order],
+  );
+
+  if (sort_order - Math.max(...sortOrderArr) > 1) {
+    return {
+      queryError: "sort_order is too large",
+    };
+  }
+
+  const SQL = `
+    UPDATE choices SET sort_order = $1 
+    WHERE id = $2
+    RETURNING *;
+  `;
+
+  const { rows } = await pbPool.query(SQL, [sort_order, choice_id]);
+
+  if (existingChoice.rowCount > 0) {
+    await pbPool.query("UPDATE choices set sort_order = $1 WHERE id = $2", [
+      currentSort,
+      existingChoice.rows[0].id,
+    ]);
+  }
+
   return rows;
 }
 
@@ -231,11 +316,40 @@ async function deleteChoiceById(user_id, story_id, choice_id) {
   const isOwner = await storyOwnerCheck(user_id, story_id);
   if (!isOwner) return null;
 
+  const fromPassageIdResults = await pbPool.query(
+    "SELECT from_passage_id FROM choices WHERE id = $1",
+    [choice_id],
+  );
+
+  if (!fromPassageIdResults.rowCount > 0) {
+    return {
+      queryError: "choice not found",
+    };
+  }
+
+  const fromPassageId = fromPassageIdResults.rows[0].from_passage_id;
+
   const SQL = `
     DELETE FROM choices WHERE id = $1 RETURNING *
   `;
 
   const { rows } = await pbPool.query(SQL, [choice_id]);
+
+  const sortOrderArr = (
+    await pbPool.query(
+      "SELECT id, sort_order FROM choices WHERE from_passage_id = $1 ORDER BY sort_order ASC",
+      [fromPassageId],
+    )
+  ).rows;
+
+  if (sortOrderArr.length > 0) {
+    for (let i = 0; i < sortOrderArr.length; i++) {
+      await pbPool.query("UPDATE choices SET sort_order = $1 WHERE id = $2", [
+        i,
+        sortOrderArr[i].id,
+      ]);
+    }
+  }
   return rows;
 }
 
@@ -249,6 +363,7 @@ export default {
   getChoicesByStoryId,
   getChoiceById,
   insertNewChoice,
+  updateChoiceSortOrderById,
   updateChoiceById,
   deleteChoiceById,
 };
